Refactor graph and unit modules into modular packages
