+++ coggraph/__init__.py
@@ -1,0 +1,5 @@
+"""CogGraph package exposing the main interfaces."""
+
+from .core import CogGraph, INPUT_CHANNELS
+
+__all__ = ["CogGraph", "INPUT_CHANNELS"]
