+++ cogunit/__init__.py
@@ -1,0 +1,5 @@
+"""CogUnit package exposing the main `CogUnit` class."""
+
+from .core import CogUnit
+
+__all__ = ["CogUnit"]
