+++ immune_cog_graph/__init__.py
@@ -1,0 +1,5 @@
+"""ImmuneCogGraph package exposing core interfaces."""
+
+from .core import ImmuneCogGraph
+
+__all__ = ["ImmuneCogGraph"]
